pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "timeseries-council"
version = "1.0.0"
description = "AI Council for Time Series Analysis - Multi-model forecasting and anomaly detection with LLM deliberation"
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
authors = [
    { name = "Manpreet Singh" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies - includes foundation models available on PyPI
#
# MODELS INCLUDED (work out of box):
#   - Chronos (Amazon) - chronos-forecasting
#   - Merlion detectors (Salesforce) - salesforce-merlion
#   - Statistical baseline - statsmodels
#
# MODELS REQUIRING SPECIAL INSTALL:
#   - Moirai2 (Python 3.10+): pip install git+https://github.com/SalesforceAIResearch/uni2ts.git
#   - TimesFM: pip install timesfm (requires JAX)
#   - Lag-Llama: pip install lag-llama (from git)
#
# Without special installs, TimesFM/Lag-Llama/Moirai2 use statistical fallbacks.
#
dependencies = [
    # Base requirements
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "statsmodels>=0.14.0",
    # PyTorch (required by all foundation models)
    "torch>=2.0.0",
    # HuggingFace ecosystem (for model downloading)
    "huggingface_hub>=0.20.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    # Amazon Chronos (works out of box!)
    "chronos-forecasting>=1.0.0",
    # Einops (used by many models)
    "einops>=0.7.0",
    # GluonTS (for probabilistic forecasting)
    "gluonts>=0.14.0",
    # Merlion (Salesforce time series library - works out of box!)
    "salesforce-merlion>=2.0.0",
    # Scikit-learn (for ML anomaly detectors)
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
# PyOD outlier detectors (ecod, copod, hbos, knn, ocsvm, loda) — needs numba/llvmlite
pyod = [
    "pyod>=1.1.0",
]
# LLM providers
llm = [
    "google-genai>=0.3.0",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
]
# Web interface
web = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
]
# Moirai2 (requires Python 3.10+) - install separately:
# pip install git+https://github.com/SalesforceAIResearch/uni2ts.git
moirai = [
    "einops>=0.7.0",
]
# Skills
skills = []
# Development
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
]
# Everything (LLM providers, web UI, extras)
all = [
    "timeseries-council[llm,web,moirai,skills,dev]"
]

[project.scripts]
timeseries-council = "timeseries_council.cli.main:main"

[project.urls]
Homepage = "https://github.com/salesforce/timeseries-council"
Documentation = "https://github.com/salesforce/timeseries-council#readme"
Repository = "https://github.com/salesforce/timeseries-council"

[tool.hatch.build.targets.wheel]
packages = ["src/timeseries_council"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/data",
    "/config.yaml",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
